Nimble: don't always build library for distribution.
Recently, Nimble moved to using 'CwlPreconditionTesting' not as a
direct compilation, but through SPM. This means they have different
build settings now.

Since Nimble is always built for distribution, but 'CwlPreconditionTesting'
doesn't, the following warning is show:
`Module 'CwlPreconditionTesting' was not compiled with library evolution support; using it means binary compatibility for 'Nimble' can't be guaranteed`.

Since building for distribution is needed mostly when building binaries
it can be simply removed from the project settings and should be enabled
whenever it's built for release.