Nimble: migrate to Tuist.
Since Nimble is still evolving and we still advance it, the assumption
here was that it's best to keep it as close as possible to the current
state implementation of Nimble (as oppose to doing stuff like changing
the location of the files, raising deployment target, removing unused
stuff and so on). Exception:
1. The swiftlint target is removed, no need to maintain it if we are
not using it.
2. The deployment destinations are now only iOS and macOS. If we ever
come to need any other, since the code is adapted to it and left
unchanged in the migration - we can easily add it.
3. Not really a change, but some of the build settings were extracted
to an xcconfig file, to keep things more organized.